62.2% quantized--could be a lot worse

+++ qml/linear.cpp
@@ -22,7 +22,7 @@ void Linear::apply(F_TY *input) {
     F_TY *w_on = this->weights + i * isize;
     F_TY r = 0;
     for (int j = 0; j < isize; j++) {
-      r += w_on[j] * (input[j]);
+      r += w_on[j] * (int)(input[j]);
       tassert(!isinf(w_on[j]));
       tassert(!isinf(input[j]));
       tassert(!isbadf(r));
